Split react-form-list.spec
