+++ packages/react-impulse-form/tests/impulse-form-list/get-input.spec.ts
@@ -10,11 +10,11 @@ it("matches the type definition", ({ scope }) => {
     }),
   ])
 
-  expectTypeOf(form.getInitial).toEqualTypeOf<
+  expectTypeOf(form.getInput).toEqualTypeOf<
     (scope: Scope) => ReadonlyArray<number>
   >()
 
-  expectTypeOf(form.getElements(scope).at(0)!.getInitial).toEqualTypeOf<
+  expectTypeOf(form.getElements(scope).at(0)!.getInput).toEqualTypeOf<
     (scope: Scope) => number
   >()
 })
@@ -22,24 +22,15 @@ it("matches the type definition", ({ scope }) => {
 it("returns empty array for empty list", ({ scope }) => {
   const form = ImpulseFormList([])
 
-  expect(form.getInitial(scope)).toStrictEqual([])
+  expect(form.getInput(scope)).toStrictEqual([])
 })
 
 it("returns an array of original values", ({ scope }) => {
   const form = ImpulseFormList([
-    ImpulseFormUnit(0, { initial: 3 }),
+    ImpulseFormUnit(0),
     ImpulseFormUnit(1),
-    ImpulseFormUnit(2, { initial: 4 }),
+    ImpulseFormUnit(2),
   ])
 
-  expect(form.getInitial(scope)).toStrictEqual([3, 1, 4])
-})
-
-it("returns nested list's values", ({ scope }) => {
-  const form = ImpulseFormList([
-    ImpulseFormList([ImpulseFormUnit(1)]),
-    ImpulseFormList([ImpulseFormUnit(2), ImpulseFormUnit(3)]),
-  ])
-
-  expect(form.getInitial(scope)).toStrictEqual([[1], [2, 3]])
+  expect(form.getInput(scope)).toStrictEqual([0, 1, 2])
 })
